Bisect FX node asserts on `ValidationException`.
This PR introduces binary search for finding smaller validation errors, when they occur.

We do that by bisecting the sequence of `torch._assert` FX nodes recorded as the source
expression of the translation validator (TV) by `ShapeEnv.evaluate_expr` calls. Then, we
raise the error caused by the earliest node.

In summary, the changes are:
- Call `bisect` on `ValidationError` @ _torch/_dynamo/convert_frame.py_
- Store a reference of the list of `TrackedFake` to `ShapeEnv` @ _torch/_dynamo/output_graph.py_
- Implement the event recording logic @ _torch/fx/experimental/symbolic_shapes.py_
    - Create the `ShapeEnvEvent` class
    - Decorate `ShapeEnv` methods with the `record_shapeenv_event` decorator
- Implement the `ShapeEnv` reconstruction function @ _torch/fx/experimental/symbolic_shapes.py_
- Implement the binary search @ _torch/fx/experimental/symbolic_shapes.py_

ghstack-source-id: 232a9fd4c09810c0fb00ea60b3ce9165424c7f89
Pull Request resolved: https://github.com/pytorch/pytorch/pull/107493