[memory snapshots] _record_memory_history_legacy bug fix
The argment order for the legacy path got swapped in a recent patch.
Because there is still a blog post documenting the legacy interface
people are hitting this pathway.

This patch fixes #108208
I will also update the blog post to the new API so that people are
more likely to use the newer `_record_memory_history` API.

ghstack-source-id: 3af8be1124dff59eb164cd5fc22111574969ffd3
Pull Request resolved: https://github.com/pytorch/pytorch/pull/108260